Parsing tmTheme (plist) themes
