--- backend/imageContainer/test/README.md
@@ -1,22 +1,0 @@
-# Markdown Theme Kit
-
-A set of themes based on [SublimeText-Markdown/MarkdownEditing](https://github.com/SublimeText-Markdown/MarkdownEditing).
-
-## Markdown Editor
-
-![](https://raw.githubusercontent.com/Microsoft/vscode-themes/master/markdown-editing/images/markdown-editor-preview.png)
-
-## Markdown Editor Dark
-
-![](https://raw.githubusercontent.com/Microsoft/vscode-themes/master/markdown-editing/images/markdown-editor-dark-preview.png)
-
-## Markdown Editor Focus
-
-![](https://raw.githubusercontent.com/Microsoft/vscode-themes/master/markdown-editing/images/markdown-editor-focus-preview.png)
-
-## Markdown Editor Yellow
-
-![](https://raw.githubusercontent.com/Microsoft/vscode-themes/master/markdown-editing/images/markdown-editor-yellow-preview.png)
-
-# LICENSE
-[MIT](LICENSE.txt)
